requirements.txt
kagglehub
tensorflow
scikit-learn
minio
numpy
pangres
pillow
streamlit